split evaluator into modules
